delete: 불필요한 파일 삭제 및 폴더 구조 변경
