+++ src/pages/main/Main.jsx
@@ -1,16 +1,16 @@
-import Contact from './Contact'
-import Intro from './Intro'
-import Work from './Work'
-import ItemsVer3 from './ItemsVer3'
-import InfoVer1 from './InfoVer1'
+import Contact from './sections/Contact'
+import Intro from './sections/Intro'
+import Work from './sections/Work'
+import Items from './sections/Items'
+import Info from './sections/Info'
 
 export default function Main() {
   return (
     <div className='w-full h-full flex flex-col gap-[150px]'>
       <Intro />
-      <InfoVer1 />
+      <Info />
       <Work />
-      <ItemsVer3 />
+      <Items />
       <Contact />
     </div>
   )
