+++ src/pages/main/sections/Contact.jsx
@@ -1,6 +1,6 @@
-import contact from '../../images/contact.jpeg'
-import contact1 from '../../images/contact1.png'
-import contact2 from '../../images/contact2.png'
+import contact from '../../../images/contact.jpeg'
+import contact1 from '../../../images/contact1.png'
+import contact2 from '../../../images/contact2.png'
 
 export default function Contact() {
   return (
